requirements.txt
fastapi
uvicorn
diffusers
transformers
accelerate
clip-anytorch
pillow
opencv-python
numpy
pydantic
python-multipart
mysql-connector-python
pymysql
redis
onnx
onnxruntime
python-dotenv
python-dotenv
